% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lms2.test.R
\name{Lms2.test}
\alias{Lms2.test}
\title{Logarithmic Minimum Test for Independence in Two-Way Contingency Table}
\usage{
Lms2.test(nij, B = 10000)
}
\arguments{
\item{nij}{a numeric matrix with non-negative values of the two-way contingency table cells}

\item{B}{an integer specifying the number of replicates used in the Monte Carlo test (optional)}
}
\value{
The function returns values of the test statistic and p-value of the logarithmic minimum test.
}
\description{
Calculates the test statistic and p-value of the logarithmic minimum test.
}
\details{
The test statistic and p-value of he logarithmic minimum test for independence in r x c contingency table,
}
\examples{
pij=matrix(1/6, nrow = 2, ncol = 3)
\donttest{Lms2.test(GenTab2(pij, 50), B = 1e3)}
iris$size <- ifelse(iris$Sepal.Length < median(iris$Sepal.Length),"small", "big")
\donttest{Lms2.test(table(iris$Species, iris$size), B=1e3)}

}
\references{
{Sulewski, P. (2019). \emph{The LMS for Testing Independence in Two-way Contingency Tables.} Biometrical Letters 56(1), 17-43}
}
\author{
Piotr Sulewski, \email{piotr.sulewski@apsl.edu.pl}, Pomeranian University in Slupsk.
}
