% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod3.pvalue.R
\name{Mod3.pvalue}
\alias{Mod3.pvalue}
\title{Modular test for independence in three-way contingency table}
\usage{
Mod3.pvalue(stat, nr, nc, nt, n, B = 10000)
}
\arguments{
\item{stat}{a modular statistic value}

\item{nr}{a number of rows}

\item{nc}{a number of columns}

\item{nt}{a number of tubes}

\item{n}{a sample size}

\item{B}{\if{html}{\out{<optional>}} an integer specifying the number of replicates used in the Monte Carlo test}
}
\description{
Calculates the p-value of the modular test for independence in three-way contingency table
}
\details{
The Critical value of the modular test for independence in r x c x t contingency table,
}
\examples{
tab7 = GenTab3(array(0.125, dim = c(2, 2, 2)), 100)
Mod3.pvalue(Mod3.stat(tab7), 2, 2, 2, 100, B=1e3)
Mod3.pvalue(1.333, 2, 3, 4, 200, B=6*1e3)

}
\references{
{Sulewski, P. (2018). \emph{Power Analysis Of Independence Testing for the Three-Way Con-tingency Tables of Small Sizes.} Journal of Applied Statistics 45(13), 2481-2498}
}
\author{
Piotr Sulewski, \email{piotr.sulewski@apsl.edu.pl}, Pomeranian University in Slupsk.
}
