% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod4.stat.R
\name{Mod4.stat}
\alias{Mod4.stat}
\title{Modular test for independence in four-way contingency table}
\usage{
Mod4.stat(nijtu)
}
\arguments{
\item{nijtu}{a numeric matrix with non-negative values of the four-way contingency table cells}
}
\value{
The function returns the value of the modular test statistic.
}
\description{
Calculates the statistic of the modular test for independence in four-way contingency table
}
\details{
The statistic of Logarithmic minimum test for independence in r x c x t x u contingency table,
}
\examples{
Mod4.stat(GenTab4(array(1/16, dim = c(2, 2, 2, 2)), 100))
Mod4.stat(table6)

}
\references{
Extension of the information contained in {Sulewski, P. (2018). \emph{Power Analysis Of Independence Testing for the Three-Way Con-tingency Tables of Small Sizes.} Journal of Applied Statistics 45(13), 2481-2498}
}
\author{
Piotr Sulewski, \email{piotr.sulewski@apsl.edu.pl}, Pomeranian University in Slupsk.
}
