% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSIndependenceTest.R
\name{PSIndependenceTest}
\alias{PSIndependenceTest}
\alias{PSIndependenceTest-package}
\title{The list of package functions and their demonstration}
\description{
The \strong{PSIndependenceTest} package puts into practice the modular and logarithmic minimum tests
for independence in two-way, three-way and four-way contingency tables. Statistic value, cv value and p-value are calculated.
This package also includes three table generation functions and six data sets.
The list of package functions is as follows:
}
\section{Data sets in the package and generating two-way, three-way and four-way contingemcy tables}{


\code{\link{table1}}

\code{\link{table2}}

\code{\link{table3}}

\code{\link{table4}}

\code{\link{table5}}

\code{\link{table6}}

\code{\link{GenTab2}}

\code{\link{GenTab3}}

\code{\link{GenTab4}}
}

\section{Functions for the modular independence test and two-way contingency table}{


\code{\link{Mod2.stat}}

\code{\link{Mod2.cv}}

\code{\link{Mod2.pvalue}}

\code{\link{Mod2.test}}
}

\section{Functions for the modular independence test and three-way contingency table}{


\code{\link{Mod3.stat}}

\code{\link{Mod3.cv}}

\code{\link{Mod3.pvalue}}

\code{\link{Mod3.test}}
}

\section{Functions for the modular independence test and four-way contingency table}{


\code{\link{Mod4.stat}}

\code{\link{Mod4.cv}}

\code{\link{Mod4.pvalue}}

\code{\link{Mod4.test}}
}

\section{Functions for the logarithmic minimum independence test and two-way contingency table}{


\code{\link{Lms2.stat}}

\code{\link{Lms2.cv}}

\code{\link{Lms2.pvalue}}

\code{\link{Lms2.test}}
}

\section{Functions for the logarithmic minimum independence test and three-way contingency table}{


\code{\link{Lms3.stat}}

\code{\link{Lms3.cv}}

\code{\link{Lms3.pvalue}}

\code{\link{Lms3.test}}
}

\section{Functions for the logarithmic minimum independence test and four-way contingency table}{


\code{\link{Lms4.stat}}

\code{\link{Lms4.cv}}

\code{\link{Lms4.pvalue}}

\code{\link{Lms4.test}}
}

