% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\docType{data}
\name{table1}
\alias{table1}
\title{First data set as two-way contingency table 2 x 2}
\format{
two-way contingency table 2 x 2
}
\usage{
table1
}
\description{
The first data set from {Sulewski, P. (2017) \emph{A new test for independence in 2x2 contingency tables,} Acta Universitatis Lodziensis. Folia Oeconomica, 4(330), 55–75}
consist of 40 observations described the effect of a treatment for rheumatoid arthritis vs. a placebo. See Table 17 in the paper.
}
\keyword{data}
\keyword{set}
