% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.R
\docType{data}
\name{table2}
\alias{table2}
\title{Second data set as two-way contingency table 2 x 3}
\format{
two-way contingency table 2 x 3
}
\usage{
table2
}
\description{
The second data set from {Sulewski, P. (2019). \emph{The LMS for Testing Independence in Two-way Contingency Tables.} Biometrical Letters 56 (1), 17-43}
consist of 25 observations described the effect of a treatment for rheumatoid arthritis vs. a placebo. See Table 17 in the paper.
}
\keyword{data}
\keyword{set}
