% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3.R
\docType{data}
\name{table3}
\alias{table3}
\title{Third data set: three-way contingency table 3 x 3 x 2}
\format{
three-way contingency table 3 x 3 x 2
}
\usage{
table3
}
\description{
The third data set from {Sulewski, P. (2021). \emph{Logarithmic Minimum Test for Independence in Three Way Con-tingency Table of Small Sizes.} Journal of Statistical Computation and Simulation 91(13), 2780-2799}
consist of 695 observations described the frequency of watching videos at home or at friends’ homes for young people between 7 and 15 years of age, cross-classified according to age and sex.
See Table 10 in the paper.
}
\keyword{data}
\keyword{set}
