% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table5.R
\docType{data}
\name{table5}
\alias{table5}
\title{Fifth data set: four-way contingency table 4 x 2 x 2 x 2}
\format{
four-way contingency table 4 x 2 x 2 x 2
}
\usage{
table5
}
\description{
The fifth data set provides information on the fate of 2201 passengers on the fatal maiden voyage of the ocean liner ‘Titanic’, summarized according to economic status (class), sex, age and survival.
}
\keyword{data}
\keyword{set}
