% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6.R
\docType{data}
\name{table6}
\alias{table6}
\title{Sixth data set: four-way contingency table 2 x 2 x 2 x 2}
\format{
four-way contingency table 2 x 2 x 2 x 2
}
\usage{
table6
}
\description{
The sixth data set obtained using the Monte Carlo method consist of 160 observations
when Ho is true, i.e. all probabilities equal 1/16.
}
\keyword{data}
\keyword{set}
