% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.R
\docType{data}
\name{table2}
\alias{table2}
\title{Second data set as two-way contingency table 2 x 3}
\format{
two-way contingency table 2 x 3
}
\usage{
table2
}
\description{
The second data set obtained using the Monte Carlo method consist of 60 observations
when Ho is true, i.e. all probabilities equal 1/6
}
\keyword{data}
\keyword{set}
