% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table4.R
\docType{data}
\name{table4}
\alias{table4}
\title{Fourth data set: three-way contingency table 2 x 2 x 2}
\format{
three-way contingency table 2 x 2 x 2
}
\usage{
table4
}
\description{
The fourth data set obtained using the Monte Carlo method consist of 80 observations
when Ho is true, i.e. all probabilities equal 1/8.
}
\keyword{data}
\keyword{set}
